Only register with Ember.libraries once.
This initializer will run once for every application boot. This is fine
for non-testing scenarios, but when testing each acceptance test will
fire the initializer resulting in the library being reregistered many
times.

When a library is re-registered the following warning is printed to the
console:

```
Library "MyAppName" is already registered with Ember.
```